requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# 
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
docx2txt>=0.8

# 
numpy>=1.24.0
scikit-learn>=1.3.0

# 
streamlit>=1.28.0
requests>=2.31.0

# orchestration
langgraph>=0.2.50
